% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeatureUMAP.R
\name{plotFeatureUMAP}
\alias{plotFeatureUMAP}
\title{Taking a SingleCellExperiment object with UMAP dimension reduction 
embeddings, make a UMAP plot colored by categorical clustering result.}
\usage{
plotFeatureUMAP(
  sce,
  feat_name = "ADIPOQ",
  facet = NULL,
  dim_red_name = "UMAP.HARMONY"
)
}
\arguments{
\item{sce}{A SingleCellExperiment (SCE) object with reduced dimension UMAPs 
in colData.}

\item{feat_name}{The name of the feature to plot, which can be any one in the 
rownames of the SCE object, e.g. `"ADIPOQ"`.}

\item{facet}{a string to specify the facet, e.g. `"orig.ident`.}

\item{dim_red_name}{default is harmony batch corrected embeddings
`"UMAP.HARMONY"`, where the first two dimensions will be plotted on x and y 
axis.}
}
\value{
a UMAP plot colored by continuous feature expression level.
}
\description{
Taking a SingleCellExperiment object with UMAP dimension reduction 
embeddings, make a UMAP plot colored by categorical clustering result.
}
\examples{
\dontrun{
## Visium DLPFC Example
library(BayesSpace)
sce <- getRDS("2020_maynard_prefrontal-cortex", "151673")
sce <- 
set.seed(123)
sce <- spatialPreprocess(sce, n.PCs = 15)
set.seed(123)
sce <- runUMAP(sce, dimred = "PCA")
p <- plotFeatureUMAP(sce, feat_name = "ENSG00000238009", facet = NULL,
                          dim_red_name = "UMAP")
p                         
}
}
